+++ src/components/molecules/FormFields/advancedSettings.module.css
@@ -1,0 +1,3 @@
+.advancedBtn {
+  margin-bottom: 2rem;
+}
