+++ src/components/pages/Publish/FormPublish.module.css
@@ -11,6 +11,3 @@
   justify-content: space-between;
   align-items: center;
 }
-.advancedBtn {
-  margin-bottom: 2rem;
-}
